Added feature search by pedal for preset params
Using the mod function, we can quickly find the correct feature_id to
corrolate with the respective pedal_id.

When requesting the feature_id, we get the global feature_id out of the
entire list.  the mod feature allows us to filter out based on how many
features per pedal there are.